remove exceptions from lite.text_generation
