+++ lite/valor_lite/text_generation/llm_clients.py
@@ -1,3 +1,4 @@
+import json
 from functools import wraps
 from typing import Any
 
@@ -11,7 +12,6 @@ try:
 except ImportError:
     OpenAI = None
 
-from valor_lite.text_generation.exceptions import InvalidLLMResponseError
 from valor_lite.text_generation.llm_instructions_analysis import (
     generate_answer_correctness_verdicts_instruction,
     generate_answer_relevance_verdicts_instruction,
@@ -27,9 +27,16 @@ from valor_lite.text_generation.llm_instructions_analysis import (
     generate_summary_coherence_instruction,
     generate_toxicity_verdicts_instruction,
 )
-from valor_lite.text_generation.utilities import trim_and_load_json
 
 DEFAULT_SYSTEM_PROMPT = "You are a helpful assistant."
+
+
+class InvalidLLMResponseError(Exception):
+    """
+    Raised when the response from the LLM is invalid for a given metric computation.
+    """
+
+    pass
 
 
 def validate_messages(messages: list[dict[str, str]]):
@@ -57,6 +64,38 @@ def validate_messages(messages: list[dict[str, str]]):
         raise ValueError("All roles in messages must be strings.")
     if not all(isinstance(message["content"], str) for message in messages):
         raise ValueError("All content in messages must be strings.")
+
+
+def trim_and_load_json(input_string: str) -> Any:
+    """
+    Trims and loads input_string as a json. Adapted from DeepEval https://github.com/confident-ai/deepeval/blob/dc117a5ea2160dbb61909c537908a41f7da4dfe7/deepeval/metrics/utils.py#L50
+
+    Parameters
+    ----------
+    input_string : str
+        The input string to trim and load as a json.
+
+    Returns
+    -------
+    Any
+        The json object.
+    """
+    start = input_string.find("{")
+    end = input_string.rfind("}") + 1
+
+    if end == 0 and start != -1:
+        input_string = input_string + "}"
+        end = len(input_string)
+
+    jsonStr = input_string[start:end] if start != -1 and end != 0 else ""
+
+    try:
+        return json.loads(jsonStr)
+    except json.JSONDecodeError as e:
+        raise InvalidLLMResponseError(
+            "Evaluation LLM outputted an invalid JSON. Please use a better evaluation model. JSONDecodeError: "
+            + str(e)
+        )
 
 
 def retry_if_invalid_llm_response():
