+++ lite/valor_lite/text_generation/utilities.py
@@ -1,7 +1,3 @@
-import json
-from typing import Any
-
-from valor_lite.text_generation.exceptions import InvalidLLMResponseError
 from valor_lite.text_generation.metric import MetricType
 
 
@@ -52,35 +48,3 @@ def validate_metric_parameters(
                 )
             if sum(bleu_weights) != 1:
                 raise ValueError("BLEU metric weights must sum to 1.")
-
-
-def trim_and_load_json(input_string: str) -> Any:
-    """
-    Trims and loads input_string as a json. Adapted from DeepEval https://github.com/confident-ai/deepeval/blob/dc117a5ea2160dbb61909c537908a41f7da4dfe7/deepeval/metrics/utils.py#L50
-
-    Parameters
-    ----------
-    input_string : str
-        The input string to trim and load as a json.
-
-    Returns
-    -------
-    Any
-        The json object.
-    """
-    start = input_string.find("{")
-    end = input_string.rfind("}") + 1
-
-    if end == 0 and start != -1:
-        input_string = input_string + "}"
-        end = len(input_string)
-
-    jsonStr = input_string[start:end] if start != -1 and end != 0 else ""
-
-    try:
-        return json.loads(jsonStr)
-    except json.JSONDecodeError as e:
-        raise InvalidLLMResponseError(
-            "Evaluation LLM outputted an invalid JSON. Please use a better evaluation model. JSONDecodeError: "
-            + str(e)
-        )
